+++ examples/erniesage/config/erniesage_linkpredict.yaml
@@ -1,27 +1,29 @@
 # Global Enviroment Settings 
 #
 # trainer config ------
-learner_type: "cpu"
+learner_type: "gpu"
 optimizer_type: "adam"
 lr: 0.00005
-batch_size: 2
+batch_size: 32
 CPU_NUM: 10
 epoch: 20
 log_per_step: 1
-save_per_step: 100
+save_per_step: 1000
 output_path: "./output"
 ckpt_path: "./ernie_base_ckpt"
 
 # data config ------
-input_data: "./data.txt"
-graph_path: "./workdir"
+train_data: "./data.txt"
+graph_data: "./data.txt"
+
+graph_work_path: "./workdir"
 sample_workers: 1
 use_pyreader: true
 input_type: "text"
 
 # model config ------
 samples: [10]
-model_type: "ErnieSageModelV1"
+model_type: "ERNIESageV3"
 layer_type: "graphsage_sum"
 
 max_seqlen: 40
@@ -31,7 +33,8 @@ hidden_size: 128
 final_fc: true
 final_l2_norm: true
 loss_type: "hinge"
-margin: 0.3
+margin: 0.1
+neg_type: "batch_neg"
 
 # infer config ------
 infer_model: "./output/last"
@@ -44,11 +47,11 @@ ernie_config:
     attention_probs_dropout_prob: 0.1
     hidden_act: "relu"
     hidden_dropout_prob: 0.1
-    hidden_size: 768
+    hidden_size: 128
     initializer_range: 0.02
     max_position_embeddings: 513
-    num_attention_heads: 12
-    num_hidden_layers: 12
+    num_attention_heads: 2
+    num_hidden_layers: 2
     sent_type_vocab_size: 4
     task_type_vocab_size: 3
     vocab_size: 18000
