+++ examples/erniesage/dataset/graph_reader.py
@@ -74,17 +74,15 @@ class GraphGenerator(BaseDataGenerator):
         batch_dst = np.array(batch_dst, dtype="int64")
 
         if self.neg_type == "batch_neg":
-            neg_shape = [1]
+            batch_neg = batch_dst
         else:
+            # TODO user define shape of neg_sample
             neg_shape = batch_dst.shape
-        sampled_batch_neg = alias_sample(neg_shape, self.alias, self.events)
-    
-        if len(batch_neg) > 0:
+            sampled_batch_neg = alias_sample(neg_shape, self.alias, self.events)
             batch_neg = np.concatenate([batch_neg, sampled_batch_neg], 0)
-        else:
-            batch_neg = sampled_batch_neg
 
         if self.phase == "train":
+            # TODO user define ignore edges or not
             #ignore_edges = np.concatenate([np.stack([batch_src, batch_dst], 1), np.stack([batch_dst, batch_src], 1)], 0)
             ignore_edges = set()
         else:
@@ -92,7 +90,8 @@ class GraphGenerator(BaseDataGenerator):
 
         nodes = np.unique(np.concatenate([batch_src, batch_dst, batch_neg], 0))
         subgraphs = graphsage_sample(self.graph, nodes, self.samples, ignore_edges=ignore_edges)
-        #subgraphs[0].reindex_to_parrent_nodes(subgraphs[0].nodes)
+        subgraphs[0].node_feat["index"] = subgraphs[0].reindex_to_parrent_nodes(subgraphs[0].nodes).astype(np.int64)
+        subgraphs[0].node_feat["term_ids"] = self.term_ids[subgraphs[0].node_feat["index"]].astype(np.int64)
         feed_dict = {}
         for i in range(self.num_layers):
             feed_dict.update(self.graph_wrappers[i].to_feed(subgraphs[i]))
@@ -103,9 +102,12 @@ class GraphGenerator(BaseDataGenerator):
         sub_neg_idx = subgraphs[0].reindex_from_parrent_nodes(batch_neg)
 
         feed_dict["user_index"] = np.array(sub_src_idx, dtype="int64")
-        feed_dict["item_index"] = np.array(sub_dst_idx, dtype="int64")
+        feed_dict["pos_item_index"] = np.array(sub_dst_idx, dtype="int64")
         feed_dict["neg_item_index"] = np.array(sub_neg_idx, dtype="int64")
-        feed_dict["term_ids"] = self.term_ids[subgraphs[0].node_feat["index"]].astype(np.int64)
+
+        feed_dict["user_real_index"] = np.array(sub_src_idx, dtype="int64")
+        feed_dict["pos_item_real_index"] = np.array(sub_dst_idx, dtype="int64")
+        feed_dict["neg_item_real_index"] = np.array(sub_neg_idx, dtype="int64")
         return feed_dict
 
     def __call__(self):
